Create Lowest_Common_Ancestor_I.py
Given two nodes in a binary tree, find their lowest common ancestor.
Examples

        5

      /   \

     9     12

   /  \      \

  2    3      14

The lowest common ancestor of 2 and 14 is 5

The lowest common ancestor of 2 and 9 is 9

way of thinking:
None, if the subtree does not include any of p or q.
p, if the subtree only contains p.
q, if the subtree only contains q.
if the subtree contains both p and q, then it should return the lowest common ancestor (i.e. it is in fact the actual answer the cod need return).